// src/models/modelsMovie.tsx
export  interface MovieModels {

    _id : string,
    rank : string;
    title : string;
    description: string;
    image: string;
    big_image: string;
    genre : string[];
    thumbnail: string;
    rating : number;
    id: string;
    year: number;
    imdbid: string;
    imdb_link: string;



} 